[TELE-1725] Correctly handle query parameter with string of array as value
The array should not be parsed and split into multiple query params
as there is a difference of having the same parameter multiple times
and having a parameter once but with a string as a value. Even
if that string contains the JSON representation of an array.

Authors: @JDurstberger @jordanrobinson